% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_.R
\name{arrange_colData}
\alias{arrange_colData}
\alias{arrange_colData.MultiAssayExperiment}
\alias{arrange_colData.SummarizedExperiment}
\title{Arrange columns of a \code{MultiAssayExperiment} or \code{SummarizedExperiment}}
\usage{
arrange_colData(.data, ...)

\method{arrange_colData}{MultiAssayExperiment}(.data, ...)

\method{arrange_colData}{SummarizedExperiment}(.data, ...)
}
\arguments{
\item{.data}{Either a \code{MultiAssayExperiment} or \code{SummarizedExperiment}
object.}
}
\description{
Allow columns of assays within a \code{MultiAssayExperiment} or
\code{SummarizedExperiment} to be rearranged by the observations of their
\code{colData}.
}
