% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_.R
\name{arrange_rowData}
\alias{arrange_rowData}
\alias{arrange_rowData.MultiAssayExperiment}
\alias{arrange_rowData.SummarizedExperiment}
\title{Arrange rows of a \code{MultiAssayExperiment} or \code{SummarizedExperiment}}
\usage{
arrange_rowData(.data, ...)

\method{arrange_rowData}{MultiAssayExperiment}(.data, ...)

\method{arrange_rowData}{SummarizedExperiment}(.data, ...)
}
\arguments{
\item{.data}{Either a \code{MultiAssayExperiment} or \code{SummarizedExperiment}
object.}
}
\description{
Arrange rows of a \code{MultiAssayExperiment} or \code{SummarizedExperiment}
}
