% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_.R
\name{filter_colData}
\alias{filter_colData}
\alias{filter_colData.MultiAssayExperiment}
\alias{filter_colData.SummarizedExperiment}
\title{Subset columns of \code{SummarizedExperiment}s by rows of \code{colData}.}
\usage{
filter_colData(.data, ...)

\method{filter_colData}{MultiAssayExperiment}(.data, ...)

\method{filter_colData}{SummarizedExperiment}(.data, ...)
}
\arguments{
\item{.data}{A \code{SummarizedExperiment} or \code{MultiAssayExperiment}.}

\item{...}{<\code{\link[dplyr]{dplyr_data_masking}}> Expressions that return
a logical value, and are defined in terms of the variables in \code{.data}. If
multiple expressions are included, they are combined with the \code{&} operator.
Only rows for which all conditions evaluate to \code{TRUE} are kept.}
}
\description{
Subsets \code{colData}, retaining only rows that produce a value of
\code{TRUE} for all conditions. The columns of \code{assay} data are then subset to
match.
}
