% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_.R
\name{filter_rowData}
\alias{filter_rowData}
\alias{filter_rowData.MultiAssayExperiment}
\alias{filter_rowData.SummarizedExperiment}
\title{Subset columns of \code{SummarizedExperiment}s by rows of \code{rowData}.}
\usage{
filter_rowData(.data, ...)

\method{filter_rowData}{MultiAssayExperiment}(.data, ...)

\method{filter_rowData}{SummarizedExperiment}(.data, ...)
}
\arguments{
\item{.data}{A \code{SummarizedExperiment} or \code{MultiAssayExperiment}.}

\item{...}{Expressions that return
a logical value, and are defined in terms of the variables in \code{.data}. If
multiple expressions are included, they are combined with the \code{&} operator.
Only rows for which all conditions evaluate to \code{TRUE} are kept.}
}
\description{
Subset columns of \code{SummarizedExperiment}s by rows of \code{rowData}.
}
