% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_colData.R
\name{intersect_colData}
\alias{intersect_colData}
\title{Keep biological units based on matching factors in \code{colData}}
\usage{
intersect_colData(.data, experiment1, experiment2, by)
}
\arguments{
\item{.data}{A MultiAssayExperiment object.}

\item{experiment1}{A SummarizedExperiment within \code{.data}.}

\item{experiment2}{Another SummarizedExperiment within \code{.data}.}

\item{by}{A vector of \code{colData} features (i.e. column names) to match units
by.}
}
\description{
\code{MultiAssayExperiment::intersectColumns()} only works if the
experiments have the same IDs within the \code{sampleMap}. This function will
use \code{colData} features to match biological units (currently only supports
two experiments).
}
