% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{mae}
\alias{mae}
\title{A \code{MultiAssayExperiment} Object}
\format{
An object of class \code{MultiAssayExperiment} of length 2.
}
\usage{
mae
}
\description{
26 patients were randomly assigned to a \code{treatment} or
\code{placebo}.

On each of 5 visits, stool samples were obtained for 16S rRNA-seq (of 10
bacterial species). However, 5 of the sequencing results failed quality
control and had to be discarded. The results are stored in the \code{counts}
assay in the \code{phy16S} \code{SummarizedExperiment}.

On visits 1, 3, and 5, the concentrations of 8 cytokines were measured.
These results are stored in the \code{cyto_conc} assay in the \code{cyto}
\code{SummarizedExperiment}.

(The data within this \code{MultiAssayExperiment} is entirely fabricated. It
exists as a "toy" example of a \code{MultiAssayExperiment}.)
}
\examples{
print(mae)
}
\keyword{datasets}
