% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_colData.R
\name{map_colData}
\alias{map_colData}
\alias{map_colData.MultiAssayExperiment}
\title{Map global \code{colData} to local}
\usage{
map_colData(.data)

\method{map_colData}{MultiAssayExperiment}(.data)
}
\arguments{
\item{.data}{A MultiAssayExperiment object.}
}
\description{
Ensures that each Experiment within a MultiAssayExperiment has
\code{colData} relevant to its samples.
}
