% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_assistants.R
\name{scale_rowwise}
\alias{scale_rowwise}
\title{Scaling and Centering of Rows in Matrix-like Objects}
\usage{
scale_rowwise(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{a numeric matrix(like object), such as an assay.}

\item{center}{a logical value indicating whether rows should be centered.}

\item{scale}{a logical value indicating whether rows should be scaled.}
}
\description{
Extends \code{base::scale} to center and/or scale the rows of a numeric matrix.
}
