% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_.R
\name{arrange_columns}
\alias{arrange_columns}
\alias{arrange_columns.MultiAssayExperiment}
\alias{arrange_columns.SummarizedExperiment}
\title{Arrange columns of a \code{MultiAssayExperiment} or \code{SummarizedExperiment}}
\usage{
arrange_columns(.data, by1, by2)

\method{arrange_columns}{MultiAssayExperiment}(.data, by1, by2)

\method{arrange_columns}{SummarizedExperiment}(.data, by1, by2)
}
\arguments{
\item{.data}{Either a \code{MultiAssayExperiment} or \code{SummarizedExperiment} object.}
}
\description{
Allow columns of assays within a \code{MultiAssayExperiment} or \code{SummarizedExperiment} to be rearranged by the observations of their \code{colData}.
}
