% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_.R
\name{arrange_rowData}
\alias{arrange_rowData}
\alias{arrange_rowData.MultiAssayExperiment}
\alias{arrange_rowData.SummarizedExperiment}
\title{Arrange rows of a \code{MultiAssayExperiment} or \code{SummarizedExperiment}}
\usage{
arrange_rowData(.data, experiment, ...)

\method{arrange_rowData}{MultiAssayExperiment}(.data, experiment, ...)

\method{arrange_rowData}{SummarizedExperiment}(.data, ...)
}
\arguments{
\item{.data}{Either a \code{MultiAssayExperiment} or \code{SummarizedExperiment}
object.}

\item{...}{Columns in \code{rowData} to arrange by.}

\item{`experiment`}{If \code{.data} is a \code{MultiAssayExperiment}, the experiment to
operate upon must be specified.}
}
\value{
An object of the same class as \code{.data}.
}
\description{
Allow rows of assays within a \code{MultiAssayExperiment} or
\code{SummarizedExperiment} to be rearranged by the observations of their
\code{rowData}. Unlike \code{colData}, \code{rowData} is experiment-specific, so only a
single experiment can be operated on at a time.
}
