% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_colData.R
\name{intersect_colData}
\alias{intersect_colData}
\title{Keep biological units based on matching factors in \code{colData}}
\usage{
intersect_colData(.data, by)
}
\arguments{
\item{.data}{A \code{MultiAssayExperiment} object.}

\item{by}{A vector of \code{colData} features (i.e. column names) to match units
by. The column names must appear in the \code{colData} for each experiment.}
}
\value{
A \code{MultiAssayExperiment} whose experiments all contain \code{colData} with
matching features.
}
\description{
\code{\link[MultiAssayExperiment:MultiAssayExperiment-helpers]{MultiAssayExperiment::intersectColumns()}} only works if the
experiments have the same IDs within the \code{sampleMap}. This function will
use \code{colData} features to match biological units.
}
\examples{
intersect_colData(mae, by = c("Patient", "Visit"))

}
