% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-transmute.R
\name{mutate}
\alias{mutate}
\title{Create, modify, and delete assays}
\usage{
mutate(.data, ...)
}
\description{
\code{mutate()} adds new assays and preserves existing ones. New
assays overwrite existing assays of the same name.
}
