% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{MultiAssayExperiment}
\alias{experiments}
\alias{SummarizedExperiment}
\alias{assays}
\alias{assays<-}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}}

  \item{MultiAssayExperiment}{\code{\link[MultiAssayExperiment:MultiAssayExperiment-methods]{experiments}}, \code{\link[MultiAssayExperiment]{MultiAssayExperiment}}}

  \item{SummarizedExperiment}{\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assays}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{assays<-}}, \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment}}}
}}

