% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-transmute.R, R/reexports.R,
%   R/select-pull.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{mutate}
\alias{transmute}
\alias{\%>\%}
\alias{MultiAssayExperiment}
\alias{experiments}
\alias{SummarizedExperiment}
\alias{assays}
\alias{assays<-}
\alias{select}
\alias{pull}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{select}}, \code{\link[dplyr:mutate]{transmute}}}

  \item{MultiAssayExperiment}{\code{\link[MultiAssayExperiment:MultiAssayExperiment-methods]{experiments}}, \code{\link[MultiAssayExperiment]{MultiAssayExperiment}}}

  \item{SummarizedExperiment}{\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assays}}, \code{\link[SummarizedExperiment:SummarizedExperiment-class]{assays<-}}, \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment}}}
}}

