% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-pull.R
\name{select}
\alias{select}
\alias{select.MultiAssayExperiment}
\title{Select \code{SummarizedExperiment}s from a \code{MultiAssayExperiment}}
\usage{
select(.data, ...)

\method{select}{MultiAssayExperiment}(.data, ...)
}
\arguments{
\item{.data}{A \code{MultiAssayExperiment}.}

\item{...}{The names of the \code{SummarizedExperiment}s to be kept.}
}
\value{
A \code{MultiAssayExperiment} containing the selected \code{SummarizedExperiment}s.
}
\description{

}
