% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-pull.R
\name{select}
\alias{select}
\alias{select.MultiAssayExperiment}
\title{Subset a \code{MultiAssayExperiment} by \code{SummarizedExperiment}s.}
\usage{
select(.data, ...)

\method{select}{MultiAssayExperiment}(.data, ...)
}
\arguments{
\item{.data}{A \code{MultiAssayExperiment}.}

\item{...}{The names of the \code{SummarizedExperiment}s to be kept, unquoted.}
}
\value{
A \code{MultiAssayExperiment} containing the selected
\code{SummarizedExperiment}s. The output has the following properties:
\itemize{\item \code{colData}, \code{rowData}, and \code{assay}s of selected experiments
are not affected. \item Output experiments are a subset of input
experiments, potentially with a different order.}
}
\description{

}
