% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-pull.R
\name{select}
\alias{select}
\title{Subset a \code{MultiAssayExperiment} by experiments.}
\arguments{
\item{.data}{A \code{MultiAssayExperiment}.}

\item{...}{The names of the experiments to be kept, unquoted.}
}
\value{
A \code{MultiAssayExperiment} containing the selected
\code{SummarizedExperiment}s. The output has the following properties:
\itemize{\item Attributes of selected experiments (e.g. \code{colData},
\code{rowData}, and \code{assay}s for \code{SummarizedExperiment} objects) are not
affected. \item Output experiments are a subset of input experiments,
potentially with a different order.}
}
\description{
Select experiments in a \code{MultiAssayExperiment}.
}
\examples{
# Return a MultiAssayExperiment object containing only 
# the SummarizedExperiment cyto
mae \%>\% select(cyto)

}
\seealso{
\code{\link[dplyr:select]{dplyr::select()}}
}
