% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_empty_.R
\name{trim_empty_rows}
\alias{trim_empty_rows}
\alias{trim_empty_rows.MultiAssayExperiment}
\alias{trim_empty_rows.SummarizedExperiment}
\title{Remove rows containing all zeroes}
\usage{
trim_empty_rows(.data, experiment, counts = TRUE)

\method{trim_empty_rows}{MultiAssayExperiment}(.data, experiment, counts = TRUE)

\method{trim_empty_rows}{SummarizedExperiment}(.data, counts = TRUE)
}
\arguments{
\item{.data}{Either a \code{MultiAssayExperiment} or a \code{SummarizedExperiment}.}

\item{experiment}{If \code{.data} is a \code{MultiAssayExperiment}, choose which
experiment to remove missing values from.}

\item{counts}{If \code{TRUE}, empty rows can be checked using \code{rowSums(x) > 0},
which is faster than \code{any(x != 0)}.}
}
\value{
An object of the same class as \code{.data}.
}
\description{
Sequencing data often produces large but sparse matrices. To
save space and time, it may be worthwhile to remove empty columns and/or
rows. Note that a row will only be removed if it is empty for all assays in
a \code{SummarizedExperiment}.
}
